Lambda_Function_as_image/insights_generator/requirements.txt
boto3
pdfkit
pandas
google-generativeai
opencv-contrib-python
python-dotenv
markdown
openpyxl
